requirements.txt
# Web framework
fastapi>=0.115
uvicorn[standard]>=0.30

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
psycopg2-binary>=2.9

# LLM / LangChain
langchain>=1.0
langchain-openai>=1.0
langchain-core>=1.0
langchain-community>=0.3
langgraph>=1.0
langgraph-prebuilt>=1.0
openai>=1.50

# Tavily search
tavily-python>=0.3

# HTTP (weather tool + test client)
httpx>=0.27

# Config
pydantic-settings>=2.5

# Observability
opentelemetry-sdk>=1.25

# Testing
pytest>=8.0
pytest-asyncio>=0.23
aiosqlite>=0.19
respx>=0.20
